# specs/001-distributed-betting-ledger/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Iron Ledger — Wallet Operations API
  version: 1.0.0
  description: |
    Moves player wallet balances for betting operations. Every endpoint here is backed by the
    same application use case as the SQS consumer, so the financial guarantees documented in
    `spec.md` (FR-003) hold identically on both channels.

    **Authentication**: OAuth 2.0 client_credentials against Keycloak. Send the resulting
    access token as `Authorization: Bearer <token>`. This is a machine-to-machine API; there
    are no end-user sessions and no login endpoint.

    **Money**: every monetary field is an integer in the currency's minimum unit plus an ISO
    4217 code. `amountMinor: 3000, currency: "BRL"` is thirty reais. There is no float, no
    decimal string, and no JSON `number` for money anywhere in this contract. Values with
    excess precision are rejected, never rounded.

    **Tenancy**: one Keycloak realm per tenant. The `iss` claim identifies the tenant and is
    validated against the issuer configured for that deployment. A token from another tenant
    is rejected with `403 TENANT_MISMATCH`. Every response carries the tenant that owns it.

    **Idempotency**: every mutating endpoint requires `Idempotency-Key`. Replaying the same
    key returns the original stored response verbatim with `X-Idempotent-Replay: true`. A
    replay with a *different* body is rejected with `409 IDEMPOTENCY_KEY_CONFLICT`.
servers:
  - url: https://ledger.internal
...
security:
  - bearerAuth: []
...
paths:
  /v1/wallets/{walletId}/operations:
    post:
      operationId: applyOperation
      tags: [Operations]
      summary: Apply a financial operation to a wallet
      description: |
        Single entry point for all five operation types. Dispatches on `type` after
        validating the request, the currency and the wallet status.

        Guarantees, all enforced by the database rather than by this service:

        - **No negative balance** — a debit that would take the balance below zero is
          rejected with `409 INSUFFICIENT_FUNDS` and writes nothing.
        - **Exactly once** — the same `Idempotency-Key` never produces a second financial
          effect, including after a full restart of every instance.
        - **Append-only** — a correction is a new compensating entry; the referenced entry is
          never modified or removed.
        - **Reconciliation** — the reported balance always equals the sum of the wallet's
          ledger entries.
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRequest'
            examples:
              bet:
                summary: Place a bet (debits the wallet)
                value:
                  type: BET
                  amountMinor: 3000
                  currency: BRL
                  transactionId: txn-8f21a0
                  referenceId: bet-9001
              win:
                summary: Settle a bet as a win (credits the prize)
                value:
                  type: WIN
                  betId: 3f2a1b4c-5d6e-4f70-8a91-b2c3d4e5f607
                  amountMinor: 9000
                  currency: BRL
                  transactionId: txn-8f21b1
              loss:
                summary: Settle a bet as a loss (no balance movement)
                value:
                  type: LOSS
                  betId: 3f2a1b4c-5d6e-4f70-8a91-b2c3d4e5f607
                  amountMinor: 3000
                  currency: BRL
                  transactionId: txn-8f21b2
              refund:
                summary: Refund a previous debit (credits it back)
                value:
                  type: REFUND
                  betId: 3f2a1b4c-5d6e-4f70-8a91-b2c3d4e5f607
                  amountMinor: 3000
                  currency: BRL
                  transactionId: txn-8f21c3
              rollback:
                summary: Reverse an operation with compensating entries
                value:
                  type: ROLLBACK
                  operationId: 7c6d5e4f-3a2b-4c1d-9e8f-7a6b5c4d3e2f
                  amountMinor: 9000
                  currency: BRL
                  transactionId: txn-8f21c4
              usdBet:
                summary: A USD wallet accepts only USD
                value:
                  type: BET
                  amountMinor: 5000
                  currency: USD
                  transactionId: txn-8f21c5
                  referenceId: bet-9002
      responses:
        '200':
          description: |
            The operation was applied, or it was already applied and the stored original
            result is being replayed. Check `X-Idempotent-Replay` to tell the two apart.
          headers:
            X-Idempotent-Replay:
              description: '`true` when this is a replay of a previously completed operation.'
              schema: { type: boolean }
            Location:
              description: URL of the resulting bet, for BET.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResult'
        '400':
          description: Malformed request — unparseable amount, unknown operation type, missing field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                badAmount:
                  value:
                    error: { code: INVALID_AMOUNT, message: 'amountMinor must be a positive integer; precision finer than the currency minor unit is not accepted', field: amountMinor }
        '401':
          description: Missing, malformed or unverifiable access token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: |
            Token is valid but not permitted — unknown `azp`, missing required scope, or the
            token's realm does not match this deployment's tenant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: No such wallet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: |
            The request is well-formed but violates a financial invariant. Nothing was
            written. The `code` distinguishes every case.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                insufficientFunds:
                  summary: The debit would make the balance negative
                  value:
                    error: { code: INSUFFICIENT_FUNDS, message: 'wallet balance 2000 would become -1000', availableMinor: 2000 }
                duplicateKey:
                  summary: Same idempotency key, different request body
                  value:
                    error: { code: IDEMPOTENCY_KEY_CONFLICT, message: 'this idempotency key was used with a different request body' }
                betAlreadySettled:
                  value:
                    error: { code: BET_ALREADY_SETTLED, message: 'bet 3f2a1b4c... is already SETTLED_WIN' }
                betNotFound:
                  value:
                    error: { code: BET_NOT_FOUND, message: 'no bet 3f2a1b4c... for this wallet' }
                currencyMismatch:
                  value:
                    error: { code: CURRENCY_MISMATCH, message: 'operation currency BRL does not match wallet currency USD', walletCurrency: USD }
                walletFrozen:
                  value:
                    error: { code: WALLET_FROZEN, message: 'wallet is FROZEN' }
                insufficientBalanceForRollback:
                  value:
                    error: { code: INSUFFICIENT_FUNDS, message: 'rollback would require a negative balance; deposit funds first', availableMinor: 500 }
        '429':
          description: Rate limited by the edge. Retry with backoff; the operation is idempotent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: Dependency unavailable — the request was not applied. Safe to retry with the same key.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/wallets/{walletId}:
    get:
      operationId: getWallet
      tags: [Wallet]
      summary: Read the current balance
      description: |
        The returned balance is materialized for speed but is always exactly equal to
        `SUM(direction * amount_minor)` over the wallet's ledger entries.
      parameters:
        - $ref: '#/components/parameters/WalletId'
      responses:
        '200':
          description: Current balance and status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Keycloak access token from the `client_credentials` grant. Validated locally against
        the realm JWKS: signature (`RS256` only), `iss`, `aud`, `exp`. Authorization is then
        derived from `azp` and `scope`.

        **Deployment prerequisite**: the audience mapper must be configured on this API's
        client so that `aud` contains this API's client id. Without it a valid token is
        rejected — see `quickstart.md`.
  parameters:
    WalletId:
      name: walletId
      in: path
      required: true
      description: Wallet identifier.
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-chosen key, unique per wallet, stable across retries. Max 255 characters.
        Required on every mutating request — a request without one is rejected with
        `400 MISSING_IDEMPOTENCY_KEY` rather than being processed unprotected.
      schema: { type: string, minLength: 1, maxLength: 255 }
  responses:
    Unauthorized:
      description: Missing, malformed or unverifiable access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token is valid but not permitted for this tenant or operation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such wallet, or it belongs to another tenant.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Currency:
      type: string
      description: |
        ISO 4217 code. Supported in this release: BRL and USD, both with 2 decimal places.
        There is no implicit conversion — an operation must be in the wallet's own currency.
      enum: [BRL, USD]
      example: BRL
    MinorUnits:
      type: integer
      format: int64
      description: |
        Integer amount in the currency's minimum unit. For BRL and USD, 1 unit = 0.01.
        Always an integer on the wire: JSON `number` with a fractional part is invalid.
      minimum: 0
      example: 3000
...
    OperationType:
      type: string
      enum: [BET, WIN, LOSS, REFUND, ROLLBACK]
    OperationRequest:
      type: object
      required: [type, amountMinor, currency, transactionId]
      additionalProperties: false
      properties:
        type:
          $ref: '#/components/schemas/OperationType'
        amountMinor:
          allOf: [{ $ref: '#/components/schemas/MinorUnits' }]
          minimum: 1
          description: |
            Positive integer strictly greater than zero. `BET`, `WIN`, `REFUND` and
            `ROLLBACK` move this amount. For `LOSS` it states the stake being settled and no
            balance movement occurs; it is still required and still validated, so a LOSS with
            an amount that does not match the bet's stake is detectable.
        currency: { $ref: '#/components/schemas/Currency' }
        transactionId:
          type: string
          minLength: 1
          maxLength: 128
          description: |
            The caller's business transaction reference, unique per wallet. A second distinct
            operation reusing a `transactionId` is rejected with `409 DUPLICATE_TRANSACTION_ID`.
        betId:
          type: string
          format: uuid
          description: |
            Required for `WIN`, `LOSS`, `REFUND` and `ROLLBACK`. Omitted for `BET`.
        operationId:
          type: string
          format: uuid
          description: Required for `ROLLBACK` — the operation being reversed.
        referenceId:
          type: string
          maxLength: 128
          description: |
            Optional external identifier for a `BET`, unique per tenant. Lets the caller find
            the created bet without reading the response.
    OperationResult:
      type: object
      required: [operationId, type, status, balanceMinor, currency, ledgerEntryIds, occurredAt]
      properties:
        operationId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/OperationType' }
        status:
          type: string
          enum: [COMPLETED]
        betId:
          type: [string, 'null']
          format: uuid
        balanceMinor: { $ref: '#/components/schemas/MinorUnits' }
        currency: { $ref: '#/components/schemas/Currency' }
        ledgerEntryIds:
          type: array
          description: |
            Entries this operation appended. `BET`, `WIN` and `REFUND` append one;
            `ROLLBACK` may append two (a credit and a debit, or vice versa); `LOSS` appends
            none, because settling a loss moves no money.
          items: { type: string, format: uuid }
        reversedOperationId:
          type: [string, 'null']
          format: uuid
          description: Set on a `ROLLBACK` result — the operation it reversed.
        occurredAt: { type: string, format: date-time }
...
    Wallet:
      type: object
      required: [id, currency, balanceMinor, status]
      properties:
        id: { type: string, format: uuid }
        playerId: { type: string }
        currency: { $ref: '#/components/schemas/Currency' }
        balanceMinor: { $ref: '#/components/schemas/MinorUnits' }
        status: { type: string, enum: [ACTIVE, FROZEN] }
        version: { type: integer, format: int64 }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable reason. Callers branch on this, not on the message.
              enum:
                - MISSING_IDEMPOTENCY_KEY
                - INVALID_AMOUNT
                - INVALID_CURRENCY
                - CURRENCY_MISMATCH
                - UNKNOWN_OPERATION_TYPE
                - UNKNOWN_FIELD
                - MISSING_FIELD
                - WALLET_NOT_FOUND
                - WALLET_FROZEN
                - BET_NOT_FOUND
                - INSUFFICIENT_FUNDS
                - BET_ALREADY_SETTLED
                - INVALID_OPERATION_STATE
                - IDEMPOTENCY_KEY_CONFLICT
                - DUPLICATE_TRANSACTION_ID
                - REFERENCE_ALREADY_EXISTS
                - UNAUTHORIZED
                - FORBIDDEN
                - TENANT_MISMATCH
                - RATE_LIMITED
                - DEPENDENCY_UNAVAILABLE
                - INTERNAL_ERROR
            message: { type: string, description: Human-readable. Not stable — do not branch on it. }
            field: { type: [string, 'null'], description: 'Offending request field, when applicable.' }
            availableMinor:
              type: [integer, 'null']
              format: int64
              description: Current available balance on `INSUFFICIENT_FUNDS`, so the caller can size the failure without a second request.
            walletCurrency: { type: [string, 'null'], enum: [BRL, USD, null] }
        requestId: { type: string, description: Correlates the response with server logs and the audit record. }